+++ src/main/java/com/nhnacademy/couponapi/persistence/domain/CouponPolicyBook.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.couponapi.persistance.domain;
+package com.nhnacademy.couponapi.persistence.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,31 +9,31 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "coupon_policy_category")
-public class CouponPolicyCategory {
+@Table(name = "coupon_policy_book")
+public class CouponPolicyBook {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long couponPolicyCategoryId;
+    private Long couponPolicyBookId;
 
     @ManyToOne
     @JoinColumn(name = "coupon_policy_id", nullable = false)
     private CouponPolicy couponPolicy;
 
-    @Column(name = "category_id", nullable = false)
-    private Long categoryId;
+    @Column(name = "book_id", nullable = false)
+    private Long bookId;
 
     @Builder
-    public CouponPolicyCategory(Long couponPolicyCategoryId, CouponPolicy couponPolicy, Long categoryId) {
-        this.couponPolicyCategoryId = couponPolicyCategoryId;
+    public CouponPolicyBook(Long couponPolicyBookId, CouponPolicy couponPolicy, Long bookId) {
+        this.couponPolicyBookId = couponPolicyBookId;
         this.couponPolicy = couponPolicy;
-        this.categoryId = categoryId;
+        this.bookId = bookId;
     }
 
     public void setCouponPolicy(CouponPolicy couponPolicy) {
         this.couponPolicy = couponPolicy;
     }
 
-    public void setCategoryId(Long categoryId) {
-        this.categoryId = categoryId;
+    public void setBookId(Long bookId) {
+        this.bookId = bookId;
     }
 }
