+++ src/main/java/com/nhnacademy/couponapi/persistence/domain/CouponPolicyCategory.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.couponapi.persistance.domain;
+package com.nhnacademy.couponapi.persistence.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,31 +9,31 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "coupon_policy_book")
-public class CouponPolicyBook {
+@Table(name = "coupon_policy_category")
+public class CouponPolicyCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long couponPolicyBookId;
+    private Long couponPolicyCategoryId;
 
     @ManyToOne
     @JoinColumn(name = "coupon_policy_id", nullable = false)
     private CouponPolicy couponPolicy;
 
-    @Column(name = "book_id", nullable = false)
-    private Long bookId;
+    @Column(name = "category_id", nullable = false)
+    private Long categoryId;
 
     @Builder
-    public CouponPolicyBook(Long couponPolicyBookId, CouponPolicy couponPolicy, Long bookId) {
-        this.couponPolicyBookId = couponPolicyBookId;
+    public CouponPolicyCategory(Long couponPolicyCategoryId, CouponPolicy couponPolicy, Long categoryId) {
+        this.couponPolicyCategoryId = couponPolicyCategoryId;
         this.couponPolicy = couponPolicy;
-        this.bookId = bookId;
+        this.categoryId = categoryId;
     }
 
     public void setCouponPolicy(CouponPolicy couponPolicy) {
         this.couponPolicy = couponPolicy;
     }
 
-    public void setBookId(Long bookId) {
-        this.bookId = bookId;
+    public void setCategoryId(Long categoryId) {
+        this.categoryId = categoryId;
     }
 }
