+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/CouponPolicyBookRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.CouponPolicyBook;
+import com.nhnacademy.couponapi.persistence.domain.CouponPolicyBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponPolicyBookRepository extends JpaRepository<CouponPolicyBook, Long> {
