+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/CouponPolicyCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.CouponPolicyCategory;
+import com.nhnacademy.couponapi.persistence.domain.CouponPolicyCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponPolicyCategoryRepository extends JpaRepository<CouponPolicyCategory, Long> {
