+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/CouponPolicyRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.CouponPolicy;
+import com.nhnacademy.couponapi.persistence.domain.CouponPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponPolicyRepository extends JpaRepository<CouponPolicy, Long> {
