+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/CouponPolicyTagRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.CouponPolicyTag;
+import com.nhnacademy.couponapi.persistence.domain.CouponPolicyTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponPolicyTagRepository extends JpaRepository<CouponPolicyTag, Long> {
