+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/CouponRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.Coupon;
+import com.nhnacademy.couponapi.persistence.domain.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
