+++ src/main/java/com/nhnacademy/couponapi/persistence/repository/UserCouponRepository.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.couponapi.persistance.repository;
+package com.nhnacademy.couponapi.persistence.repository;
 
-import com.nhnacademy.couponapi.persistance.domain.UserCoupon;
+import com.nhnacademy.couponapi.persistence.domain.UserCoupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserCouponRepository extends JpaRepository<UserCoupon, Long> {
